Restructure lambda functions
