+++ funcs/redirect.js
